feat(forensic): likelihood-ratio output, calibration, and evaluation metrics
Second addition in the forensic-linguistics phase. Introduces evidential-LR
output for any tamga classifier or verifier, plus the standard forensic
evaluation suite: C_llr, ECE, Brier, and Tippett-plot data.

Why this matters: forensic journals (IJSLL, Language and Law) and courtroom
gatekeeping (Daubert, R v T) expect evidence framed as a likelihood ratio with
a calibrated underlying scorer — not raw classifier posteriors, which are
easily abused as "probability of guilt" in ways that misrepresent forensic
semantics.

New modules under `tamga.forensic`:

  `tamga.forensic.lr`
    - `log_lr_from_probs(p, base=10)` — calibrated posterior → log-LR (flat
      priors).
    - `log_lr_from_probs_with_priors(p, prior_target)` — posterior → log-LR
      with user-specified prior.
    - `CalibratedScorer(method="platt"|"isotonic")` — fit a monotone
      calibrator on (scores, binary labels) then expose `predict_proba` and
      `predict_log_lr`. Platt (logistic) for small samples; isotonic for
      larger calibration sets.

  `tamga.forensic.metrics`
    - `cllr(log_lrs, y)` — Brummer & du Preez (2006) log-likelihood-ratio
      cost; the forensic proper scoring rule. Prior-only = 1.0, perfect ~ 0,
      misleading > 1.
    - `ece(probs, y, n_bins=10)` — expected calibration error.
    - `brier(probs, y)` — Brier score.
    - `tippett(log_lrs, y)` — data for a Tippett plot (target / non-target
      CDFs over log-LR thresholds).

Top-level re-exports: `CalibratedScorer`, `log_lr_from_probs`,
`log_lr_from_probs_with_priors`. Metrics stay under `tamga.forensic.metrics`
as they are forensic-specific and shouldn't pollute the tamga top-level
surface.

Tests (38 new cases, all pass):
- C_llr: prior-only = 1.0 exactly; perfect separation -> near-zero; misleading
  -> > 1; shape / class / empty validation.
- ECE: perfectly calibrated -> zero; systematically biased -> large;
  validation.
- Brier: perfect = 0; confident-wrong = 1; uninformed = 0.25.
- Tippett: monotone-decreasing CDFs; target dominates non-target on a
  discriminating system.
- log_lr_from_probs: p=0.5 -> 0; p=0.99 -> > 1; p=0.01 -> < -1; base control;
  eps clipping; validation.
- Priors: prior=0.5 matches flat; rare prior deflates log-LR.
- CalibratedScorer: contract (predict-before-fit, non-binary labels, tiny
  data, shape mismatch, unknown method); both methods beat prior-only on
  separable data; both preserve score ordering (monotone); both produce
  valid probs in [0, 1].

Drive-by: redundant `int()` around `round()` in verify.py:140 (ruff RUF046).
Swapped non-ASCII Unicode characters in metrics.py docstring for ASCII
(RUF002). Replaced a `zip(a[:-1], a[1:])` with `itertools.pairwise` (RUF007).

Full suite: 283/283 tests pass. mypy clean on src/tamga/forensic.

Stacks on PR #4 (feat/forensic-general-impostors) — merge that first.

References:
Brummer, N., & du Preez, J. (2006). Application-independent evaluation of
  speaker detection. Computer Speech & Language, 20(2-3), 230-275.
Platt, J. C. (1999). Probabilistic outputs for SVMs and comparisons to
  regularized likelihood methods. Advances in Large Margin Classifiers,
  61-74.
Niculescu-Mizil, A., & Caruana, R. (2005). Predicting good probabilities with
  supervised learning. Proceedings of ICML 2005, 625-632.

+++ src/tamga/forensic/verify.py
@@ -137,7 +137,7 @@ class GeneralImpostors:
 
         n_features = q.shape[0]
         n_impostors = i_pool.shape[0]
-        k_per_iter = max(1, int(round(self.feature_subsample_rate * n_features)))
+        k_per_iter = max(1, round(self.feature_subsample_rate * n_features))
         m_per_iter = self.impostor_sample_size or max(1, int(np.ceil(np.sqrt(n_impostors))))
         m_per_iter = min(m_per_iter, n_impostors)
